refactor to separate functions, minor rust clean up
